refactor: descohesionando el algoritmo
